initial refactoring and adding the in memory components
