+++ src/main/java/com/hivemq/persistence/ioc/PersistenceModule.java
@@ -25,8 +25,7 @@ import com.hivemq.common.shutdown.ShutdownHooks;
 import com.hivemq.configuration.service.PersistenceConfigurationService;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.mqtt.topic.tree.TopicTreeStartup;
-import com.hivemq.persistence.CleanUpService;
-import com.hivemq.persistence.PersistenceShutdownHookInstaller;
+import com.hivemq.persistence.*;
 import com.hivemq.persistence.ioc.annotation.PayloadPersistence;
 import com.hivemq.persistence.ioc.annotation.Persistence;
 import com.hivemq.persistence.ioc.provider.local.PayloadPersistenceScheduledExecutorProvider;
@@ -59,6 +58,12 @@ public class PersistenceModule extends SingletonModule<Class<PersistenceModule>>
     protected void configure() {
 
         install(new LocalPersistenceModule(persistenceInjector, persistenceConfigurationService));
+
+        if (persistenceConfigurationService.getMode() == PersistenceConfigurationService.PersistenceMode.IN_MEMORY) {
+            bind(SingleWriterService.class).to(InMemorySingleWriterImpl.class);
+        } else {
+            bind(SingleWriterService.class).to(SingleWriterServiceImpl.class);
+        }
 
         bind(ShutdownHooks.class).toInstance(persistenceInjector.getInstance(ShutdownHooks.class));
